requirements.txt
Flask
torch
torchvision
pandas
numpy
matplotlib
scikit-learn
seaborn
opencv-python
Pillow
optuna
